refactor: Extract watch.rs from cli/mod.rs + P2 fixes
- Extract watch mode to cli/watch.rs (~270 lines)
- Add HnswIndex::count_vectors() for fast stats without loading full index (P2 #52)
- Update PROJECT_CONTINUITY.md with current state

cli/mod.rs: 2167 → 1893 lines (-274)
