+++ src/cli/mod.rs
@@ -1,14 +1,13 @@
 //! CLI implementation for cq
 
 mod display;
+mod watch;
 
 use std::collections::HashSet;
 use std::path::{Path, PathBuf};
 use std::sync::atomic::{AtomicBool, AtomicUsize, Ordering};
-use std::sync::mpsc;
 use std::sync::Arc;
 use std::thread;
-use std::time::Duration;
 
 use anyhow::{bail, Context, Result};
 use clap::{Parser, Subcommand};
@@ -16,12 +15,10 @@ use colored::Colorize;
 use crossbeam_channel::{bounded, select, Receiver, Sender};
 use ignore::WalkBuilder;
 use indicatif::{ProgressBar, ProgressStyle};
-use notify::{Config, RecommendedWatcher, RecursiveMode, Watcher};
 use rayon::prelude::*;
 
 use cqs::embedder::{Embedder, Embedding};
 use cqs::hnsw::HnswIndex;
-use cqs::nl::generate_nl_description;
 use cqs::note::parse_notes;
 use cqs::parser::{Chunk, Parser as CqParser};
 use cqs::store::{ModelInfo, SearchFilter, Store};
@@ -76,7 +73,7 @@ fn setup_signal_handler() {
 }
 
 /// Check if user requested interruption via Ctrl+C
-fn check_interrupted() -> bool {
+pub(crate) fn check_interrupted() -> bool {
     INTERRUPTED.load(Ordering::SeqCst)
 }
 
@@ -228,7 +225,7 @@ pub fn run() -> Result<()> {
         Some(Commands::Watch {
             debounce,
             no_ignore,
-        }) => cmd_watch(&cli, debounce, no_ignore),
+        }) => watch::cmd_watch(&cli, debounce, no_ignore),
         Some(Commands::Serve {
             ref transport,
             ref bind,
@@ -296,7 +293,7 @@ fn apply_config_defaults(cli: &mut Cli, config: &cqs::config::Config) {
 }
 
 /// Find project root by looking for common markers
-fn find_project_root() -> PathBuf {
+pub(crate) fn find_project_root() -> PathBuf {
     let cwd = std::env::current_dir().unwrap_or_else(|_| PathBuf::from("."));
     let mut current = cwd.as_path();
 
@@ -1366,11 +1363,8 @@ fn cmd_stats(cli: &Cli) -> Result<()> {
     let stats = store.stats()?;
 
     let cq_dir = root.join(".cq");
-    let hnsw_vectors = if HnswIndex::exists(&cq_dir, "index") {
-        HnswIndex::load(&cq_dir, "index").ok().map(|h| h.len())
-    } else {
-        None
-    };
+    // Use count_vectors to avoid loading full HNSW index just for stats
+    let hnsw_vectors = HnswIndex::count_vectors(&cq_dir, "index");
 
     if cli.json {
         let json = serde_json::json!({
@@ -1533,277 +1527,6 @@ fn cmd_callees(_cli: &Cli, name: &str, json: bool) -> Result<()> {
     }
 
     Ok(())
-}
-
-/// Watch for file changes and re-index automatically
-fn cmd_watch(cli: &Cli, debounce_ms: u64, no_ignore: bool) -> Result<()> {
-    let root = find_project_root();
-    let cq_dir = root.join(".cq");
-    let index_path = cq_dir.join("index.db");
-
-    if !index_path.exists() {
-        bail!("No index found. Run 'cqs index' first.");
-    }
-
-    let parser = CqParser::new()?;
-    let supported_ext: HashSet<_> = parser.supported_extensions().iter().cloned().collect();
-
-    println!(
-        "Watching {} for changes (Ctrl+C to stop)...",
-        root.display()
-    );
-    println!(
-        "Code extensions: {}",
-        supported_ext.iter().cloned().collect::<Vec<_>>().join(", ")
-    );
-    println!("Also watching: docs/notes.toml");
-
-    let (tx, rx) = mpsc::channel();
-
-    let config = Config::default().with_poll_interval(Duration::from_millis(debounce_ms));
-
-    let mut watcher = RecommendedWatcher::new(tx, config)?;
-    watcher.watch(&root, RecursiveMode::Recursive)?;
-
-    // Track pending changes for debouncing
-    // Cap at 10k files to prevent unbounded memory growth during rapid changes
-    const MAX_PENDING_FILES: usize = 10_000;
-    let mut pending_files: HashSet<PathBuf> = HashSet::new();
-    let mut pending_notes = false; // Track if notes.toml changed
-    let mut last_event = std::time::Instant::now();
-    let debounce = Duration::from_millis(debounce_ms);
-    let notes_path = root.join("docs/notes.toml");
-
-    loop {
-        match rx.recv_timeout(Duration::from_millis(100)) {
-            Ok(Ok(event)) => {
-                for path in event.paths {
-                    // Skip .cq directory
-                    if path.starts_with(&cq_dir) {
-                        continue;
-                    }
-
-                    // Check if it's notes.toml
-                    if path == notes_path {
-                        pending_notes = true;
-                        last_event = std::time::Instant::now();
-                        continue;
-                    }
-
-                    // Skip if not a supported extension
-                    let ext = path.extension().and_then(|e| e.to_str()).unwrap_or("");
-                    if !supported_ext.contains(ext) {
-                        continue;
-                    }
-
-                    // Convert to relative path
-                    if let Ok(rel) = path.strip_prefix(&root) {
-                        if pending_files.len() < MAX_PENDING_FILES {
-                            pending_files.insert(rel.to_path_buf());
-                        }
-                        last_event = std::time::Instant::now();
-                    }
-                }
-            }
-            Ok(Err(e)) => {
-                if !cli.quiet {
-                    eprintln!("Watch error: {}", e);
-                }
-            }
-            Err(mpsc::RecvTimeoutError::Timeout) => {
-                // Check if we should process pending changes
-                let should_process = (!pending_files.is_empty() || pending_notes)
-                    && last_event.elapsed() >= debounce;
-
-                if should_process {
-                    // Reindex code files if any changed
-                    if !pending_files.is_empty() {
-                        let files: Vec<PathBuf> = pending_files.drain().collect();
-                        if !cli.quiet {
-                            println!("\n{} file(s) changed, reindexing...", files.len());
-                            for f in &files {
-                                println!("  {}", f.display());
-                            }
-                        }
-
-                        match reindex_files(
-                            &root,
-                            &index_path,
-                            &files,
-                            &parser,
-                            no_ignore,
-                            cli.quiet,
-                        ) {
-                            Ok(count) => {
-                                if !cli.quiet {
-                                    println!("Indexed {} chunk(s)", count);
-                                }
-                            }
-                            Err(e) => {
-                                eprintln!("Reindex error: {}", e);
-                            }
-                        }
-                    }
-
-                    // Reindex notes if notes.toml changed
-                    if pending_notes {
-                        pending_notes = false;
-                        if !cli.quiet {
-                            println!("\nNotes changed, reindexing...");
-                        }
-                        match reindex_notes(&root, &index_path, cli.quiet) {
-                            Ok(count) => {
-                                if !cli.quiet {
-                                    println!("Indexed {} note(s)", count);
-                                }
-                            }
-                            Err(e) => {
-                                eprintln!("Notes reindex error: {}", e);
-                            }
-                        }
-                    }
-                }
-            }
-            Err(mpsc::RecvTimeoutError::Disconnected) => {
-                bail!(
-                    "File watcher disconnected unexpectedly. \
-                     Hint: Restart 'cqs watch' to resume monitoring."
-                );
-            }
-        }
-
-        if check_interrupted() {
-            println!("\nStopping watch...");
-            break;
-        }
-    }
-
-    Ok(())
-}
-
-/// Reindex specific files
-fn reindex_files(
-    root: &Path,
-    index_path: &Path,
-    files: &[PathBuf],
-    parser: &CqParser,
-    _no_ignore: bool,
-    quiet: bool,
-) -> Result<usize> {
-    let embedder = Embedder::new()?;
-    let store = Store::open(index_path)?;
-
-    // Parse the changed files
-    let chunks: Vec<_> = files
-        .iter()
-        .flat_map(|rel_path| {
-            let abs_path = root.join(rel_path);
-            if !abs_path.exists() {
-                // File was deleted, we'll handle this by removing old chunks
-                return vec![];
-            }
-            match parser.parse_file(&abs_path) {
-                Ok(mut file_chunks) => {
-                    // Rewrite paths to be relative
-                    for chunk in &mut file_chunks {
-                        chunk.file = rel_path.clone();
-                    }
-                    file_chunks
-                }
-                Err(e) => {
-                    tracing::warn!("Failed to parse {}: {}", abs_path.display(), e);
-                    vec![]
-                }
-            }
-        })
-        .collect();
-
-    if chunks.is_empty() {
-        return Ok(0);
-    }
-
-    // Generate embeddings with neutral sentiment for code chunks
-    let texts: Vec<String> = chunks.iter().map(generate_nl_description).collect();
-    let text_refs: Vec<&str> = texts.iter().map(|s| s.as_str()).collect();
-    let embeddings: Vec<Embedding> = embedder
-        .embed_documents(&text_refs)?
-        .into_iter()
-        .map(|e| e.with_sentiment(0.0))
-        .collect();
-
-    // Delete old chunks for these files and insert new ones
-    for rel_path in files {
-        store.delete_by_origin(rel_path)?;
-    }
-
-    for (chunk, embedding) in chunks.iter().zip(embeddings.iter()) {
-        let abs_path = root.join(&chunk.file);
-        let mtime = abs_path
-            .metadata()
-            .and_then(|m| m.modified())
-            .ok()
-            .and_then(|t| t.duration_since(std::time::UNIX_EPOCH).ok())
-            .map(|d| d.as_secs() as i64);
-        store.upsert_chunk(chunk, embedding, mtime)?;
-    }
-
-    if !quiet {
-        println!("Updated {} file(s)", files.len());
-    }
-
-    Ok(chunks.len())
-}
-
-/// Reindex notes from docs/notes.toml
-fn reindex_notes(root: &Path, index_path: &Path, quiet: bool) -> Result<usize> {
-    let notes_path = root.join("docs/notes.toml");
-    if !notes_path.exists() {
-        return Ok(0);
-    }
-
-    let notes = parse_notes(&notes_path)?;
-    if notes.is_empty() {
-        return Ok(0);
-    }
-
-    let embedder = Embedder::new()?;
-    let store = Store::open(index_path)?;
-
-    // Embed note content with sentiment prefix
-    let texts: Vec<String> = notes.iter().map(|n| n.embedding_text()).collect();
-    let text_refs: Vec<&str> = texts.iter().map(|s| s.as_str()).collect();
-    let base_embeddings = embedder.embed_documents(&text_refs)?;
-
-    // Add sentiment as 769th dimension
-    let embeddings_with_sentiment: Vec<Embedding> = base_embeddings
-        .into_iter()
-        .zip(notes.iter())
-        .map(|(emb, note)| emb.with_sentiment(note.sentiment()))
-        .collect();
-
-    // Get file mtime
-    let file_mtime = notes_path
-        .metadata()
-        .and_then(|m| m.modified())
-        .ok()
-        .and_then(|t| t.duration_since(std::time::UNIX_EPOCH).ok())
-        .map(|d| d.as_secs() as i64)
-        .unwrap_or(0);
-
-    // Delete old notes and insert new
-    store.delete_notes_by_file(&notes_path)?;
-    let note_embeddings: Vec<_> = notes.into_iter().zip(embeddings_with_sentiment).collect();
-    store.upsert_notes_batch(&note_embeddings, &notes_path, file_mtime)?;
-
-    if !quiet {
-        let (total, warnings, patterns) = store.note_stats()?;
-        println!(
-            "  Notes: {} total ({} warnings, {} patterns)",
-            total, warnings, patterns
-        );
-    }
-
-    Ok(note_embeddings.len())
 }
 
 /// Start the MCP server for IDE integration
